--- src/main/java/com/gdg/homepage/security/jwt/filter/JwtAccessDeniedHandler.java
@@ -1,37 +1,0 @@
-package com.gdg.homepage.security.jwt.filter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.gdg.homepage.core.response.ApiResponse;
-import com.gdg.homepage.core.response.CustomException;
-import com.gdg.homepage.core.response.ErrorCode;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAccessDeniedHandler implements org.springframework.security.web.access.AccessDeniedHandler {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException {
-
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-        response.setContentType("application/json");
-        response.setCharacterEncoding("UTF-8");
-
-        ApiResponse<Object> fail = ApiResponse.fail(
-                new CustomException(ErrorCode.Forbidden, accessDeniedException.getMessage())
-        );
-
-        objectMapper.writeValue(response.getWriter(), fail);
-    }
-
-
-}
